# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Description removed for better illustration of structure.",
        "contact": {
            "name": "Pra feira",
            "url": "https://prafeira.com.br",
            "email": "admin@example.test"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticacao"
                ],
                "summary": "Fazer login",
                "operationId": "4b7becb0762c996662e47743a2706980",
                "requestBody": {
                    "description": "Dados de login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticacao"
                ],
                "summary": "Fazer logout",
                "operationId": "b0e3cbb0dab3aed6f3ae5edfab36ec3a",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de autenticação (Bearer)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/api/empresas": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Retorna empresa",
                "operationId": "e5d40f6ab397a2e07d02f50d3eb395f3",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Adiciona uma nova empresa",
                "operationId": "b400bda696995ef09cec01e24e345e39",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "description": "Nome da empresa",
                                        "type": "string"
                                    },
                                    "cnpj_cpf": {
                                        "description": "CNPJ da empresa",
                                        "type": "string"
                                    },
                                    "cor_tema": {
                                        "description": "Cor do tema da empresa",
                                        "type": "string"
                                    },
                                    "data_cadastro": {
                                        "description": "Data de cadastro da empresa",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "data_encerramento": {
                                        "description": "Data de encerramento da empresa",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "situacao_empresa": {
                                        "description": "Situação da empresa",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa adicionada com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/empresas/{id}": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Retorna detalhes de uma empresa específica",
                "operationId": "01cf4a33a64fd815d116b70387a19e57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da empresa a ser recuperada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Edita uma empresa existente",
                "operationId": "d6113af3348fcbc0ad3ca2ed8ce6e89e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da empresa a ser editada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "description": "Nome da empresa",
                                        "type": "string"
                                    },
                                    "cnpj_cpf": {
                                        "description": "CNPJ da empresa",
                                        "type": "string"
                                    },
                                    "cor_tema": {
                                        "description": "Cor do tema da empresa",
                                        "type": "string"
                                    },
                                    "data_cadastro": {
                                        "description": "Data de cadastro da empresa",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "data_encerramento": {
                                        "description": "Data de encerramento da empresa",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "situacao_empresa": {
                                        "description": "Situação da empresa",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa editada com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Deleta uma empresa",
                "operationId": "b8ac3362ec3de93acdd573849c335811",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da empresa a ser deletada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Empresa deletada com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}